+++ api/app/src/main/kotlin/packit/model/PacketGroupSummary.kt
@@ -6,5 +6,5 @@ interface PacketGroupSummary
     fun getName(): String
     fun getPacketCount(): Int
     fun getLatestId(): String
-    fun getLatestTime(): Long
+    fun getLatestTime(): Double
 }
